# perf: optimize getFileCommitDate, make it async