--- a/packages/docusaurus-plugin-content-docs/src/lastUpdate.ts
+++ b/packages/docusaurus-plugin-content-docs/src/lastUpdate.ts
@@ -25,10 +25,11 @@ export async function getFileLastUpdate(
   // Wrap in try/catch in case the shell commands fail
   // (e.g. project doesn't use Git, etc).
   try {
-    const result = getFileCommitDate(filePath, {
+    const result = await getFileCommitDate(filePath, {
       age: 'newest',
       includeAuthor: true,
     });
+
     return {timestamp: result.timestamp, author: result.author};
   } catch (err) {
     if (err instanceof GitNotFoundError) {