## Motivation

`getFileCommitDate()` has been identified as a strong sync IO bottleneck of the `loadContent` blog and docs lifecycle hooks. 

Making it async greatly optimizes the time it takes to call `loadContent` on all plugins. This only affects the production build because `getFileCommitDate()` was already bypassed in dev mode.

## Local benchmark

`yarn clear:website && yarn build:website --locale en`

**Before:**

- `loadContent`: 17s
- Total build time: 84s

**After:** 🔥

- `loadContent`: 1.2s
- Total build time: 68s

## Test Plan

unit tests + CI

### Test links

https://deploy-preview-9890--docusaurus-2.netlify.app/